Fixed several problems
Achieved to fix several problems on the program, but still having some timeouts inside the "Play" method, can't really debug any further with this IDE as when i try to set cerr output, its to much to be traced on the stream. Would be easy if i could use breakpoints and memory manager from Visual studio.

On the other hand, the heuristic seems not to be performing well(or maybe its something on the system that does not work) and can't figure out how to make it work, have already tested Scores, and seems to be working, simulation is working, sorting is working, hard to tell again just with cerr output.

I will try to improve the Score heuristic, but i don't think i will make it work good enough to be competitive in time.